Reimplemented the setup_fp methods
In order to resolve bug #7, we need to remove non-trivial pieces of code
by @wertarbyte, who does not react to requests currently. All other
authors agreed to any suitable OSS license. See
 https://github.com/phillipberndt/autorandr/issues/7
for further discussion.

A side-effect of my reimplementation is that both methods now return the
same fingerprint, such that they are interchangeable. A downside is that
the fingerprints change with this version.