Adhere XDG basedir spec regarding order
It says that for duplicate information (i.e., equal file names), only
the one that comes first ought to be used. So we have to make sure to
load profiles in the correct order -- the ones that have the highest
priority last, because we just dict.update them all together. And in
script execution, we have to keep a list of file names already used in
scripts that have already been executed.