wip: Support for --fb and --dpi
This is an attempt to piggy-back fb and dpi onto the first used output.
Not certain whether this has catastrophic drawbacks yet. In any case,
the dpi information should probably be obtained from the outputs, as
those already have all the necessary information, rather than by parsing
the xrandr output a second time. Maybe it'd be an even better idea to go
with --fbmm instead of dpi.

See #33