Fix case in which x and y are not specified in panning
The regular expression that extracts the panning values has optional
matches for the named groups x and y.

Previously, the case in which there is no match for x and y looks to
have been attempted to be handled by doing

  detail.get("x", "0")

This is a bug.  The default argument to dict.get is only returned when
the key is not a member of the dictionary.  However, when a group
fails to participate in the regex match, the dict does have a key.
The default value of this default key is None.  This would lead to an
error when the get returned None and None was then cast to an integer.

This resolves the issue by explicitly setting the groupdict default to
be "0", which is appropriate for the two groups (x and y) that may
optionally participate in the match.