Add an option to match profiles based on edid
In some cases the display names change so the profile matching fails.
This commit introduces the "--match-edid" option, that allows one to
specify that the profiles matching should be done based on display edid.

This is done by updating the profiles after loading, based on current
display name/edid mapping requested using xrandr.