Force autorandr on autostart
In some situations autorandr detects the screen setup as "already applied" and skip processing. But it is for the xrandr settings only.
I use autorandr mainly to change pulseaudio output settings. Forcing autorandr in autostart forces to set pulseaudio and maybe other things, so non-xrandr parts are properly set up.