Do not check for mask in evt->response_type
Value of `evt->response_type` comes from X11 server and at least in
Xorg it's not constant but depends on number of loaded extensions
https://gitlab.freedesktop.org/xorg/xserver/-/blob/21b3dad2380639573c251d95910bcb644049ee5f/mi/miinitext.c#L110
and a number of arguments each loaded extension takes. When randr
module is loaded
https://gitlab.freedesktop.org/xorg/xserver/-/blob/d6c02ffd9c910637f6b3b7249507998e9e45f93c/randr/randr.c#L439
`REventBase` is calculated.

On Debian 12 its value is 89 so the method _incidentally_ works as 89
& 1 is 1 but on my Slackware system it's value is 88 and since 88 & 1
isn't 1 the method doesn't work.

Fixes #375