Ignore processes with wrong XAUTHORITY file
This commit adds a check to see if the XAUTHORITY file exists for a
given display. If it doesn't, the process environment is added to the
backup candidates and the loop continues. This is particularly useful
for tmux processes when the user re-logs in, as the XAUTHORITY may not
be updated.

Signed-off-by: Songmin Li <lisongmin@protonmail.com>