fix: prefer serial no over serial text
The current serial based matching behaviour was specified in d9b1953.
The commit message specifies that matching on serial number should be
preferred over constructing a text based fingerprint from timing
information (which matches legacy behaviour of autorandr).

However, the actual code has the conditional reversed however, which
means that the new behaviour was never employed. This means the matching
behaviour is still suffering from the previous issues like changing edid
data when the same monitor switches interface etc (DP to HDMI for
example) or the port on the gpu changes and now has different
capabilities etc.

This commit switches the conditional around, using serial no. if it is
not zero, using the legacy serial_text if available and otherwise
setting serial to None